users authorizations added

+++ src/components/ProjectCards.js
@@ -1,4 +1,5 @@
 import { useProjectsContext } from "../hooks/useProjectsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 import { currencyFormatter } from "../utils/currencyFormatter";
 import moment from "moment";
 import { useState } from "react";
@@ -9,12 +10,20 @@ const ProjectCards = ({ project }) => {
   const [isOverlyOpen, setIsOverlyOpen] = useState(false);
 
   const { dispatch } = useProjectsContext();
+  const { user } = useAuthContext();
 
   const handleDelete = async () => {
+    if (!user) {
+      return;
+    }
+
     const res = await fetch(
       `http://localhost:5000/api/projects/${project._id}`,
       {
         method: "DELETE",
+        headers: {
+          Authorization: `Bearer ${user.token}`,
+        },
       }
     );
     const data = await res.json();
